+++ packages/abi-gen-wrappers/package.json
@@ -13,11 +13,10 @@
         "build": "yarn pre_build && tsc -b",
         "build:ci": "yarn build",
         "lint": "tslint --format stylish --project .",
-        "artifacts_transform": "node ../monorepo-scripts/lib/contract_artifacts_transform.js ../contract-artifacts/artifacts && prettier --write ../contract-artifacts/artifacts/*.json",
         "fix": "tslint --fix --format stylish --project .",
         "pre_build": "yarn generate_contract_wrappers",
         "clean": "shx rm -rf lib wrappers",
-        "generate_contract_wrappers": "yarn artifacts_transform && abi-gen --abis  ${npm_package_config_abis} --template ../../node_modules/@0x/abi-gen-templates/contract.handlebars --partials '../../node_modules/@0x/abi-gen-templates/partials/**/*.handlebars' --output src/generated-wrappers --backend ethers"
+        "generate_contract_wrappers": "abi-gen --abis  ${npm_package_config_abis} --template ../../node_modules/@0x/abi-gen-templates/contract.handlebars --partials '../../node_modules/@0x/abi-gen-templates/partials/**/*.handlebars' --output src/generated-wrappers --backend ethers"
     },
     "config": {
         "abis": "../contract-artifacts/artifacts/@(AssetProxyOwner|DutchAuction|DummyERC20Token|DummyERC721Token|ERC20Proxy|ERC20Token|ERC721Proxy|ERC721Token|Exchange|Forwarder|IAssetProxy|IValidator|IWallet|MultiAssetProxy|OrderValidator|WETH9|ZRXToken|Coordinator|CoordinatorRegistry).json"
@@ -46,7 +45,8 @@
         "shx": "^0.2.2"
     },
     "dependencies": {
-        "@0x/base-contract": "^5.1.0"
+        "@0x/base-contract": "^5.1.0",
+        "@0x/contract-artifacts": "^1.5.1"
     },
     "publishConfig": {
         "access": "public"
