+++ packages/contract-artifacts/package.json
@@ -10,8 +10,8 @@
         "test": "test"
     },
     "scripts": {
-        "artifacts_copy": "node ../monorepo-scripts/lib/contract_artifacts_copy.js ${npm_package_config_contractsPackages}",
-        "artifacts_transform": "node ../monorepo-scripts/lib/contract_artifacts_transform.js ./artifacts && prettier --write ./artifacts/*.json &&  cp -r ./artifacts/ ../../python-packages/contract_artifacts/src/zero_ex/contract_artifacts/artifacts/",
+        "artifacts_copy": "node lib/src/copy.js ${npm_package_config_contractsPackages}",
+        "artifacts_transform": "node lib/src/transform.js ./artifacts && prettier --write ./artifacts/*.json &&  cp -r ./artifacts/ ../../python-packages/contract_artifacts/src/zero_ex/contract_artifacts/artifacts/",
         "artifacts_update": "yarn artifacts_copy && yarn artifacts_transform",
         "build": "yarn tsc -b",
         "build:ci": "yarn build",
@@ -33,6 +33,7 @@
     },
     "homepage": "https://github.com/0xProject/0x-monorepo/packages/contract-artifacts/README.md",
     "devDependencies": {
+        "@0x/utils": "^4.3.3",
         "chai": "^4.0.1",
         "lodash": "^4.17.11",
         "mocha": "^4.1.0",
