+++ packages/contract-artifacts/test/contract_artifacts_test.ts
@@ -4,7 +4,8 @@ import 'mocha';
 
 import * as artifacts from '../src/index';
 
-import { ForbiddenProperties, RequiredProperties } from '../src/properties';
+import { ObjectMap } from '../../types/lib';
+import { FORBIDDEN_PROPERTIES, REQUIRE_PROPERTIES } from '../src/transform';
 
 const expect = chai.expect;
 
@@ -12,7 +13,7 @@ describe('Contract Artifacts', () => {
     it('should not include forbidden attributes', () => {
         const forbiddenPropertiesByArtifact: { [name: string]: string[] } = {};
         for (const [artifactName, artifact] of Object.entries(artifacts)) {
-            for (const forbiddenProperty of ForbiddenProperties) {
+            for (const forbiddenProperty of FORBIDDEN_PROPERTIES) {
                 const rejectedValue = get(artifact, forbiddenProperty);
                 if (rejectedValue) {
                     const previousForbidden = forbiddenPropertiesByArtifact[artifactName];
@@ -25,9 +26,9 @@ describe('Contract Artifacts', () => {
         expect(forbiddenPropertiesByArtifact).to.eql({});
     });
     it('should include all required attributes', () => {
-        const missingRequiredPropertiesByArtifact: { [name: string]: string[] } = {};
+        const missingRequiredPropertiesByArtifact: ObjectMap<string[]> = {};
         for (const [artifactName, artifact] of Object.entries(artifacts)) {
-            for (const requiredProperty of RequiredProperties) {
+            for (const requiredProperty of REQUIRE_PROPERTIES) {
                 const requiredValue = get(artifact, requiredProperty);
                 if (requiredValue === undefined || requiredValue === '') {
                     const previousMissing = missingRequiredPropertiesByArtifact[artifactName];
