+++ packages/utils/src/index.ts
@@ -1,6 +1,7 @@
 export { promisify } from './promisify';
 export { addressUtils } from './address_utils';
 export { classUtils } from './class_utils';
+export { deleteNestedProperty } from './delete_nested_property';
 export { intervalUtils } from './interval_utils';
 export { providerUtils } from './provider_utils';
 export { BigNumber } from './configured_bignumber';
